requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.4.1
colorlog==6.7.0
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.35
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.4
imageio==2.31.3
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.24.3
pygame==2.5.1
pyparsing==3.0.9
pytorch-lightning==2.0.8
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
scikit-learn==1.3.0
scipy==1.11.2
sentry-sdk==1.30.0
setproctitle==1.3.2
smmap==5.0.0
submitit==1.4.6
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.1
torchmetrics==1.1.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
urllib3==2.0.4
wandb==0.15.10
yarl==1.9.2
